refactor: update user adapter test folder structure
