+++ test/adapters/controllers/user/delete-user-by-id-controller.spec.ts
@@ -1,22 +1,22 @@
-import { makeUserRepository } from '@/test/stubs';
-import { ReadUserUseCase } from '@/use-cases/read-user';
-import { NonExistingUserError } from '@/use-cases/errors';
+import { makeFakeRequestAuthenticated, makeUserRepository } from '@/test/stubs';
+import { NonExistingUserError } from '@/use-cases/user/errors';
 import { ServerError } from '@/adapters/errors';
 import { serverError, unauthorized } from '@/adapters/util/http';
 import { UserBuilder } from '@/test/builders/user-builder';
 import { error, success } from '@/shared';
-import { ReadUserController } from '@/adapters/controllers/read-user-controller';
+import { DeleteUserByIdController } from '@/adapters/controllers/user/delete-user-by-id-controller';
+import { DeleteUserUseCase } from '@/use-cases/user/delete-user';
 
 type SutTypes = {
-  sut: ReadUserController,
-  useCase: ReadUserUseCase,
+  sut: DeleteUserByIdController,
+  useCase: DeleteUserUseCase,
   user: UserBuilder
 };
 
 const makeSut = (): SutTypes => {
   const repository = makeUserRepository();
-  const useCase = new ReadUserUseCase(repository);
-  const sut = new ReadUserController(useCase);
+  const useCase = new DeleteUserUseCase(repository);
+  const sut = new DeleteUserByIdController(useCase);
   const user = new UserBuilder();
 
   return {
@@ -26,16 +26,15 @@ const makeSut = (): SutTypes => {
   };
 };
 
-describe('ReadUser Controller ', () => {
-  it('Should call ReadUserUseCase with correct values', async () => {
+describe('DeleteUser Controller ', () => {
+  it('Should call DeleteUserUseCase with correct values', async () => {
     const { sut, useCase, user } = makeSut();
 
     const useCaseSpy = jest.spyOn(useCase, 'execute');
 
     await sut.handle({
-      id: user.build().id,
-      accessToken: 'any_accessToken',
-      body: {},
+      ...makeFakeRequestAuthenticated(),
+      params: { id: user.build().id },
     });
     expect(useCaseSpy).toHaveBeenCalledWith(user.build().id);
   });
@@ -43,12 +42,13 @@ describe('ReadUser Controller ', () => {
   it('Should return 500 if throws', async () => {
     const { sut, useCase, user } = makeSut();
 
-    jest.spyOn(useCase, 'execute').mockImplementationOnce(() => { throw new Error(); });
+    jest.spyOn(useCase, 'execute').mockImplementationOnce(() => {
+      throw new Error();
+    });
 
     const response = await sut.handle({
-      id: user.build().id,
-      accessToken: 'any_accessToken',
-      body: {},
+      ...makeFakeRequestAuthenticated(),
+      params: { id: user.build().id },
     });
     expect(response).toEqual(serverError(new ServerError()));
   });
@@ -63,9 +63,8 @@ describe('ReadUser Controller ', () => {
     };
     jest.spyOn(useCase, 'execute').mockResolvedValue(success(useCaseReturn));
     const response = await sut.handle({
-      id: user.build().id,
-      accessToken: 'any_accessToken',
-      body: {},
+      ...makeFakeRequestAuthenticated(),
+      params: { id: user.build().id },
     });
 
     expect(response.statusCode).toBe(200);
@@ -76,12 +75,11 @@ describe('ReadUser Controller ', () => {
     const { sut, useCase } = makeSut();
 
     jest.spyOn(useCase, 'execute').mockResolvedValue(error(new NonExistingUserError()));
-    const response = await sut.handle({
-      id: 'invalid_id',
-      accessToken: 'any_accessToken',
-      body: {},
-    });
 
+    const response = await sut.handle({
+      ...makeFakeRequestAuthenticated(),
+      params: { id: 'any_id' },
+    });
     expect(response).toEqual(unauthorized(new NonExistingUserError()));
   });
 });
