+++ test/adapters/controllers/user/read-all-users-controller.spec.ts
@@ -1,9 +1,9 @@
 import { makeUserRepository } from '@/test/stubs';
-import { ReadAllUsersUseCase } from '@/use-cases/read-all-users';
+import { ReadAllUsersUseCase } from '@/use-cases/user/read-all-users';
 import { ServerError } from '@/adapters/errors';
 import { serverError } from '@/adapters/util/http';
 import { UserBuilder } from '@/test/builders/user-builder';
-import { ReadAllUsersController } from '@/adapters/controllers/read-all-users-controller';
+import { ReadAllUsersController } from '@/adapters/controllers/user/read-all-users-controller';
 
 type SutTypes = {
   sut: ReadAllUsersController,
