+++ test/adapters/controllers/user/read-user-controller.spec.ts
@@ -1,22 +1,22 @@
 import { makeUserRepository } from '@/test/stubs';
-import { DeleteUserUseCase } from '@/use-cases/delete-user';
-import { NonExistingUserError } from '@/use-cases/errors';
+import { NonExistingUserError } from '@/use-cases/user/errors';
 import { ServerError } from '@/adapters/errors';
 import { serverError, unauthorized } from '@/adapters/util/http';
 import { UserBuilder } from '@/test/builders/user-builder';
 import { error, success } from '@/shared';
-import { DeleteUserController } from '@/adapters/controllers/delete-user-controller';
+import { ReadUserController } from '@/adapters/controllers/user/read-user-controller';
+import { ReadUserUseCase } from '@/use-cases/user/read-user';
 
 type SutTypes = {
-  sut: DeleteUserController,
-  useCase: DeleteUserUseCase,
+  sut: ReadUserController,
+  useCase: ReadUserUseCase,
   user: UserBuilder
 };
 
 const makeSut = (): SutTypes => {
   const repository = makeUserRepository();
-  const useCase = new DeleteUserUseCase(repository);
-  const sut = new DeleteUserController(useCase);
+  const useCase = new ReadUserUseCase(repository);
+  const sut = new ReadUserController(useCase);
   const user = new UserBuilder();
 
   return {
@@ -26,8 +26,8 @@ const makeSut = (): SutTypes => {
   };
 };
 
-describe('DeleteUser Controller ', () => {
-  it('Should call DeleteUserUseCase with correct values', async () => {
+describe('ReadUser Controller ', () => {
+  it('Should call ReadUserUseCase with correct values', async () => {
     const { sut, useCase, user } = makeSut();
 
     const useCaseSpy = jest.spyOn(useCase, 'execute');
@@ -43,9 +43,7 @@ describe('DeleteUser Controller ', () => {
   it('Should return 500 if throws', async () => {
     const { sut, useCase, user } = makeSut();
 
-    jest.spyOn(useCase, 'execute').mockImplementationOnce(() => {
-      throw new Error();
-    });
+    jest.spyOn(useCase, 'execute').mockImplementationOnce(() => { throw new Error(); });
 
     const response = await sut.handle({
       id: user.build().id,
@@ -78,12 +76,12 @@ describe('DeleteUser Controller ', () => {
     const { sut, useCase } = makeSut();
 
     jest.spyOn(useCase, 'execute').mockResolvedValue(error(new NonExistingUserError()));
-
     const response = await sut.handle({
       id: 'invalid_id',
       accessToken: 'any_accessToken',
       body: {},
     });
+
     expect(response).toEqual(unauthorized(new NonExistingUserError()));
   });
 });
